+++ src/shortcodes/classnames.js
@@ -1,6 +1,0 @@
-const classnames = require( "../lib/classnames" );
-
-/**
- * @returns {Function}
- */
-module.exports.shortcode = () => classnames;